Revert "topjava forth lesson first check"
