--- src/com/unrise/webapp/exception/ExistStorageException.java
@@ -1,7 +1,0 @@
-package com.unrise.webapp.exception;
-
-public class ExistStorageException extends StorageException {
-    public ExistStorageException(String uuid) {
-        super("Resume " + uuid + " already exists", uuid);
-    }
-}
