--- src/com/unrise/webapp/exception/NotExistStorageException.java
@@ -1,7 +1,0 @@
-package com.unrise.webapp.exception;
-
-public class NotExistStorageException extends StorageException {
-    public NotExistStorageException(String uuid) {
-        super("Resume " + uuid + " does not exist", uuid);
-    }
-}
