--- src/com/unrise/webapp/exception/StorageException.java
@@ -1,14 +1,0 @@
-package com.unrise.webapp.exception;
-
-public class StorageException extends RuntimeException {
-    private final String uuid;
-
-    public StorageException(String message, String uuid) {
-        super(message);
-        this.uuid = uuid;
-    }
-
-    public String getUuid() {
-        return uuid;
-    }
-}
