+++ src/com/unrise/webapp/storage/AbstractArrayStorage.java
@@ -1,13 +1,14 @@
 package com.unrise.webapp.storage;
 
-import com.unrise.webapp.exception.ExistStorageException;
-import com.unrise.webapp.exception.NotExistStorageException;
-import com.unrise.webapp.exception.StorageException;
 import com.unrise.webapp.model.Resume;
 
 import java.util.Arrays;
 
 public abstract class AbstractArrayStorage implements Storage {
+    protected static final String NOT_IN_DB = "ERROR: Resume is not in database";
+
+    protected static final String IS_IN_DB = "ERROR: Resume is already in database";
+
     protected static final String NO_SPACE_LEFT = "ERROR: Resume database is full";
 
     protected static final int STORAGE_LIMIT = 10000;
@@ -22,7 +23,8 @@ public abstract class AbstractArrayStorage implements Storage {
 
         int foundIndex = getIndex(r.getUuid());
         if (foundIndex < 0) {
-            throw new NotExistStorageException(r.getUuid());
+            System.out.println(NOT_IN_DB);
+            return;
         }
         storage[foundIndex] = r;
     }
@@ -44,42 +46,11 @@ public abstract class AbstractArrayStorage implements Storage {
         int foundIndex = getIndex(uuid);
 
         if (foundIndex < 0) {
-            throw new NotExistStorageException(uuid);
+            System.out.println(NOT_IN_DB);
+            return null;
         }
         return storage[foundIndex];
     }
 
-    @Override
-    public void save(Resume r) {
-        if (r == null) {
-            return;
-        }
-        if (size >= storage.length) {
-            throw new StorageException(NO_SPACE_LEFT, r.getUuid());
-        }
-
-        int index = getIndex(r.getUuid());
-        if (index >= 0) {
-            throw new ExistStorageException(r.getUuid());
-        }
-
-        processSave(r, index);
-    }
-
-    @Override
-    public void delete(String uuid) {
-        int index = getIndex(uuid);
-
-        if (index < 0) {
-            throw new NotExistStorageException(uuid);
-        }
-
-        processDelete(uuid, index);
-    }
-
     protected abstract int getIndex(String uuid);
-
-    protected abstract void processDelete(String uuid, int index);
-
-    protected abstract void processSave(Resume resume, int index);
 }
