+++ src/com/unrise/webapp/storage/SortedArrayStorage.java
@@ -7,17 +7,35 @@ import java.util.Arrays;
 public class SortedArrayStorage extends AbstractArrayStorage {
 
     @Override
-    protected void processSave(Resume resume, int index) {
-        int positiveIndex = Math.abs(index);
+    public void save(Resume r) {
+        if (r == null) {
+            return;
+        }
 
-        Resume[] partAfterIndex = Arrays.copyOfRange(storage, positiveIndex - 1, size);
-        storage[positiveIndex - 1] = resume;
-        System.arraycopy(partAfterIndex, 0, storage, positiveIndex, partAfterIndex.length);
-        size++;
+        int index = getIndex(r.getUuid());
+        if (size > storage.length) {
+            System.out.println(NO_SPACE_LEFT);
+        }else if (index >= 0) {
+            System.out.println(IS_IN_DB);
+        }else {
+            index = Math.abs(index);
+
+            Resume[] partAfterIndex = Arrays.copyOfRange(storage, index - 1, size);
+            storage[index - 1] = r;
+            System.arraycopy(partAfterIndex, 0, storage, index, partAfterIndex.length);
+            size++;
+        }
     }
 
     @Override
-    protected void processDelete(String uuid, int index) {
+    public void delete(String uuid) {
+        int index = getIndex(uuid);
+
+        if (index < 0) {
+            System.out.println(NOT_IN_DB);
+            return;
+        }
+
         Resume[] partAfterIndex = Arrays.copyOfRange(storage, index + 1, size);
         System.arraycopy(partAfterIndex, 0, storage, index, partAfterIndex.length);
         storage[--size] = null;
